Change layout of some screens and composables
Squashes all development commits into one for a clean tree.

---

**Commit lists (newest to oldest):**

Change layout of PlaylistScreen()

Same as commit 04135433.

Has a slight vertical padding than RadioScreen(). Not apparent to
recognize so may be ignored.

Change layout of RadioScreen()

Adds a horizontal gap between icon and header text.

Others same as 20dbce4f.

Change layout of ArtistDetails()

Sets a smaller font for "Discography".

Removes trailing colon.

Change layout of AlbumDetails()

Adds a vertical spacer between album name and artists.

Sets a smaller font for album artists.

Change layout of settings pages

Adds vertical gap between top row and body.

Uses a fixed "magic number" as height of top row icon in order to match
the top row height of other screens like AlbumScreen(). However, it is a
rough guess so they may not be exactly identical.

Change layout of TopBarWithSearch()

Adds a horizontal gap between icon and header text for a ease look.

Replaces the default search bar (with inputField) to a concise icon
(that is, inputIcon). Usually smartphones does not have a large width,
but the icon and header text have already eaten up too much space, so
the search field may look bad for it is too short and large in height.
Also, the search field does nothing to the top page of albums, artists,
etc. It only works on search result pages and becomes meaningless on
those other pages (for it doesn't change the content in the top pages,
only leaves the search terms inside the field). Therefore, it can be
safely replaced by a button.

The button is set to a fixed height 56.dp, which is a bit larger than
the icon and header text, making them not sit next to the top and leave
some visual gaps.

There are also some problems here. For example the header text is a
little bit upper than the vertical center of icon, but it is not
appearent and maybe can be ignored.

Change layout of HomeScreen()

Tries to inherit typography style from YouTube Music. The original
design uses larger texts with less gap between them, which may cause the
layout crowded.

- Set line height of text widget of home page heading to default value
  of `headlineLarge` typography, which is a bit larger than the original
  one and keeps consistency of typography style.
- Set a wider vertical gap between page title and body (album rows).
- Use `headlineSmall` for album row headings, also remove trailing
  colons from them.

 ---

 Resolves part of #43